bug 1719 - moxi should not run when passed a bad REST/URL
With this change, if libconflate cannot contact the REST/HTTP server
at the given URL the first time, it exits the process.

A better change would be to callback the caller/libconflate-API-client
so that the program (such as moxi) can cleanup more effectively.
Since all API users (like moxi) are currently stateless, this isn't
completely horrible at the moment.

Change-Id: I06ec533e6c72dd5a2c1412df8254f48afa186412
Reviewed-on: http://review.northscale.com/1099
Reviewed-by: Dustin Sallings <dustin@spy.net>
Tested-by: Dustin Sallings <dustin@spy.net>